refactor(tools): split collaboration meta builders
